finder/pyproject.toml
[tool.poetry]
name = "finder"
version = "0.1.0"
description = ""
authors = ["Данилин Даниил <d-danilin@groupprice.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
asyncpg = "^0.29.0"
python-dotenv = "^1.0.1"
fastapi = "^0.110.0"
uvicorn = "^0.27.1"
pydantic = "^2.6.3"
asyncio = "^3.4.3"
opencv-python = "^4.9.0.80"
scikit-image = "^0.22.0"
tqdm = "^4.66.2"
torch = "^2.2.1"
torchvision = "^0.17.1"
pillow = "^10.2.0"
scikit-learn = "^1.4.1.post1"
clip = "^0.2.0"
openai-clip = "^1.0.1"

[tool.poetry.scripts]
app = "finder.main:loader"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
